+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### 1.3.1
+- Added ignore cache
+
 ### 1.3.0
 - New docs
 - Added contains match type